Separate out requests tests
This allows us to run them independently of the rest of the suite,
which should save a decent amount of time on CI
