+++ tox.ini
@@ -1,6 +1,7 @@
 [tox]
 envlist=
-    py{26,27,33,34,35,36,37,38}-{requests-,}test,
+    py{26,27,33,34,35,36,37,38}-test,
+    py{26,27,33,34,35,36,37,38}-requests,
     py{26,27}-flask,
     py{26,27,33,34,35,36,37,38}-wsgi,
     py{35,36,37}-django{18,19,110,111,20,21}-migrate1-django1,
@@ -39,7 +40,8 @@ deps=
     lint: mypy
 
 commands =
-    test: nosetests --tests=tests -sv --with-coverage --cover-package=bugsnag
+    test: nosetests -sv --with-coverage --cover-package=bugsnag
+    requests: nosetests -sv --with-coverage --cover-package=bugsnag requests/test_requests_delivery.py
     wsgi: nosetests integrations/test_wsgi.py -sv --with-coverage --cover-package=bugsnag
     flask: nosetests integrations/test_flask.py -sv --with-coverage --cover-package=bugsnag
     migrate1: python tests/fixtures/django1/manage.py migrate
